[FLINK-38451] Stop logging interrupts as ERROR
When cancelling a job, we may interrupt the KafkaCommitter. This currently leads to ERROR log "Transaction ... encountered error and data has been potentially lost."

However, that exception is expected and not leading to any data loss beyond the normal inconsistencies because of cancellation. In many cases, the commit already succeeded. Further, a job restart will lead to data being eventually committed.